refactor: enhance some code
