+++ wallet/src/main/kotlin/net/thechance/wallet/entity/PendingTransactionParams.kt
@@ -1,7 +1,5 @@
-package net.thechance.wallet.service.helper
+package net.thechance.wallet.entity
 
-import net.thechance.wallet.entity.PendingTransaction
-import net.thechance.wallet.entity.Transaction
 import net.thechance.wallet.entity.user.WalletUser
 import java.util.*
