+++ wallet/src/main/kotlin/net/thechance/wallet/service/PaymentService.kt
@@ -1,10 +1,10 @@
 package net.thechance.wallet.service
 
 import net.thechance.wallet.entity.PendingTransaction
+import net.thechance.wallet.entity.Transaction
 import net.thechance.wallet.entity.toTransaction
 import net.thechance.wallet.repository.PendingTransactionRepository
 import net.thechance.wallet.repository.TransactionRepository
-import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.util.*
@@ -22,13 +22,13 @@ class PaymentService(
         val pendingTransaction = validateAndGetTransaction(userId, transactionId)
 
         val block = blockService.getCurrentBlock()
-        val transaction = pendingTransaction.toTransaction(block)
+        val transaction = pendingTransaction.toTransaction(block, Transaction.Status.SUCCESS)
         transactionRepository.save(transaction)
         pendingTransactionRepository.deleteById(transactionId)
     }
 
     private fun validateAndGetTransaction(userId: UUID, transactionId: UUID): PendingTransaction {
-        if (transactionRepository.findByIdOrNull(transactionId) != null)
+        if (transactionRepository.existsById(transactionId))
             throw IllegalArgumentException("Transaction already processed")
 
         val pendingTransaction = pendingTransactionRepository.findById(transactionId)
